+++ tnmpa/__init__.py
@@ -1,14 +1,2 @@
-import importlib.metadata
-
-__version__ = importlib.metadata.version("tnmpa")
-
-from . import quimb_vbp
-from .message_passing_ksat import (
-    BeliefPropagation,
-    SurveyPropagation,
-    TensorBeliefPropagation,
-    TensorSurveyPropagation,
-    TwoNormBeliefPropagation,
-)
-from .models import Clause, KsatInstance
-from .walksat import WalkSAT
+from .models import KSAT, Clause
+from .solvers import *
